Add local delivery missions and expand existing functions.
Added delivery missions in-system to give people a little more reason
to explore around a system, also helps people low on funds.

Added an "Am I in any danger?" generic question to give people
a rough idea of the risk in the task. Also made the risk a little random
rather than being set in stone.

Added in a few pirate taunts so that you aren't always greeted with the
same generic attack statement.

Moved all major strings to the start of the script to make translating
easier in the long run.