Loop over screen positions directly O(n) instead of O(n log n)
This has unfortunate implications w.r.t. constness.
The m_projectedPos map now has non-const Body* keys, so that when
finding a picked body the result can be non-const.
In order to find the projected position of a const Body*, there's
a helper function find_body() which does a const cast but doesn't
let the non-const Body* escape (so it's about as safe as it can be)