Persist characters (first attempt, see commit note)
Persistent characters are chreated for every DeliverPackage advert, and are
used preferentially, meaning you'll *always* see the same pool of names
appearing again and again. Not ideal. In addition to this, the game will
probably segfault when you quit, because of the way in which things are
handled by this Lua.