Fuel balance: Adder
This is a high-thrust fighter. I've given it some really bad mileage, based
on that. Because the thrust is high, you can get a lot done manually, even
on dregs. As long as you don't let the autopilot completely run out, you're
probably going to be OK.

It's a cheap ship, meaning that new players are likely to give it a try,
perhaps as their first upgrade. It will really showcase the differences in
ships.