Fuel balance: Asp Explorer
As its name suggests, this ship has been balanced with range in mind. Fully
loaded up, it's good but not a super-ship; it will get from Earth to Pluto
and back on one tank. This makes it a better than average trader.

Where it shines, though, is on an empty load-out. With a low basic hull mass
and just the essentials, this ship can keep on going and going.