Fuel balance: Stardust
Like the E.Y.E, th Stardust is a spherical ship. Not quite *as* spherical,
but it is definitely round. I went with the same reasoning.

I didn't knock the efficiency down. Whilst very similar to the E.Y.E, this
ship has a couple of drawbacks. Its extra cargo space is entirely taken
up by the bigger hyperdrive, giving it almost exactly the same hyperspace
range. It costs half as much again, and the one thruster that is different
is less powerful.

So, for the sake of balance, I let it remain a little more efficient.