Add table raw autocompletion in LuaConsole.
It doesn't work with the names that are registered in the metatable of
the table, which means pretty much every method AFAICT, but the logic is
there and it works quite nicely with a plain table hierarchy such as:

> aa.bb.cc

To trigger the autocompletion, just press Tab, and it will complete the
last word (not the one under the cursor though) with the first match. It
is NOT alphabetical yet as Lua doesn't store table keys in any reliable
order. Press Tab again and it will replace it by the next match, and so
on. When the match pool has been exhausted, it resumes to the original
word.

If you press any other key, it will reset the completion session.