implement a new base function: math.hash_random()
hash_random can be used to robustly make a selection based on some
input data. For a lot of purposes, you want this and not a random
number generator, because RNGs are hard to use nicely, and inherently
tie together any an all code that touches them.

I make no guarantees that hash_random will not change its output in
future versions, but within a single version its output should be
reliable across platforms, and because it's a hash you can explicitly
and carefully control what inputs you wish to affect it.