render rings with premultiplied alpha blending mode
I think this is what they should be. Before the Renderer object was put
in place, the ring drawing code called glEnable(GL_BLEND) but didn't
actually set the blending mode, so it just got whatever blend mode was
active from previously executed display code.

When Renderer was done, some attempt was made to use the same blend mode
as before, but either that attempt got it wrong, or (more likely imho,
having checked the appearance as of alpha 19), some change before Renderer
was done had a knock on effect and caused the wrong blend mode to be used
for rings. Or maybe the blend mode was never right.

Either way, in #972 (from tomm), the blend mode was fixed, and
premultiplied alpha was used, but this was before Renderer.  When that
patch series was re-worked in #1357 and merged as #1278, the blend mode
change wasn't included (only atmosphere changes were included).