[valgrind] initialise Gui::Widget m_size in Widget constructor
This isn't strictly necessary, since the size of a Widget should be set
during layout. However, in practice partial layout is run at various
points (e.g., when adding a widget to a container) before size values
are actually initialised, which leads to valgrind warnings.
Better to initialise to zero, in the interests of having certain
behaviour and reducing the potential for heisenbugs.