force Align to align to integer coordinates
this works around a text rendering bug in the main menu on nvidia where the
next glyph can bleed onto the end of the current one by some fractional pixel
amount. I don't know if its a GL quirk, a missing default, a text renderer bug
or a UI bug, but this will have to do for now.