onConfigureStarSystem Lua event added
The event is fired when a new <StarSystem> is created, passing in a LuaStarSystem that can be read or altered.
added factionColour attribute to <LuaStarSystem>
added SetFactionColour and SetName functions to <LuaStarSystem>
altered SectorView::Draw3D altered to work better with changes made from Lua.
refactored System class inside <Sector> to better support the SectorView changes.
Added PopulationBrightness.ua module, demonstrating usage of the event to alter brightness of labels in SectorView based on population.