Switch to checking population to get if a system in inhabited
Work for #1641
We need to know if a system is inhabited to mark uninhabited systems as not
part of a faction on the map.
Unfortunately there's no way to determine this without a full sysgen, so
we might as well switch to using population rather than count of starports
to check this and save ourselves a bunch of messy flag twiddling code.