Cache the homeworld Sector on the first distance calculation
We no longer construct the homeworld's Sector, doing mini-system
for each system it contains, every time we do a Faction distance
calculation.
Instead We cache the Sector the first time we do a distance
calculation, and then reuse it.
I considered caching the homeworld sector when the homeworld is
set from Lua, but Sector(0,0,0) constructed at that point
appears to be invalid (possibly Sol hasn't loaded yet?)