Prevent hypertarget following selection to current system
If you're using manual hypertarget locking in the SectorView,
you aren't able to select the current system as a hypertarget
(duh!).
This changes 'hypertarget follows selection' mode, to also
leave the hypertarget alone if you select is the current
system.
This makes comparing info for your destination system to that
of your current system much less annoying in that mode.