Fix jerkiness when scrolling in Far zoom mode
'Beware the Judderman when the moon is fat!'

Turns out that I was using m_pos as the origin for when calculating
star positions when I should have been using the origin of the sector
m_pos was in.
This caused a nasty judder when crossing sector boundaries while
scrolling at any sort of angle to them.