Some faction home system improvements
Home systems should now always have a name rather than just a catalogue
number, and will always be explored.
If the home system has any population it should now be a goodly amount,
however if the system had zero population before, it'll still have zero
population.
To support all this added a Faction::IsHomeSystem method and maintain
a std::set of system paths to the homesystems.