Zooming between Near and Far now switches Info panel
If you zoom in from Far to Near and you have the Faction panel up,
it wil be swiched for the Systems panel.
If you zoom out from Near to Far and you have the Systems panel up,
it will be swtiched for the Factions panel.
The panel selection will be left alone if all other circumstances.