Update data/modules/BreakdownServicing/BreakdownServicing.lua
This is supposed to allow the player to repair their hyperdrive after it has been destroyed by a malfunction. Right now it is only replaced with Class 1 Drive. It works on my version of Pioneer. I want to activate a timer to make the player wait after they land for a couple hours to simulate time spent working on it. How to do that?

This is my first functioning function. So it is most likely clumsily written and in need of a lot of optimization. Any pointers would be extremely helpful.