Fix undercarriage icon staying up while wheels being lowered
When pressing F7 with the undercarriage up, to lower it, the icon flashes to a down arrow and then goes back to an up arrow again until the undercarriage is fully deployed. Fixed by setting the icon to down while it is being lowered too, which required querying the transition status.