remove Gui::TextEntry::onFilterKeys, which was only used by the console
With the transition to SDL2, text input for the console is now filtered
with a hack in the main event loop. In the future, it should be filtered
in the UI event dispatcher based on whether there a text-accepting widget
has the focus, and it should possibly use SDL_StartTextInput and
SDL_StopTextInput.