initialise m_showCameraName in WorldView::InitObject()
Fixes a valgrind warning about a conditional using an uninitialised value.

This is actually somewhat important; in WorldView::UpdateCameraName,
the value of m_showCameraName is checked before it's initialised, and if
it's non-null (which is entirely plausible and happened in my test run),
the code calls Remove(m_showCameraName), which then gets a bogus pointer.

I'm not sure how this hasn't bitten us before... maybe worth working out.