Set speed to the forward component when switching modes
When in Manual Mode, and pressing F5 to change to Set Speed Mode, the speed was set to the magnitude of the current one. With this patch, it's set to the projection of the speed onto the target vector (or onto the ship's forward vector if no target is set), which feels more intuitively correct.