run optipng on all PNG files to optimise and fix libPNG warnings
Recent versions of libpng (since 1.6 I think?) print a warning to
stderr when loading most of our PNGs. The warning is:

libpng warning: iCCP: known incorrect sRGB profile

and for some PNGs:

libpng warning: Interlace handling should be turned on when using
png_read_image.

This commit runs optipng on all our PNG files with this command:

  optipng -o2 -nc -nb -i 0 -strip all $IMAGE

Which strips the profile data from the image and rewrites it.
This fixes all the warnings, at the cost of generating a huge
commit and touching all our files.