Change the way we clean up station state
Clearing nextRef in onGameStart can cause ad refs to be duplicated,
because some ads may have already been created by modules that ran
before us.

So now we're clearing it in two places: at scripts load/compile for the
first game, and at game end for subsequent games.

As usual, the right way to fix this is to tear down and reload the
entire Lua state between games.