move the OpenGL info dumping code to graphics/Graphics.cpp
Yes, yes, I know people wanted to separate the renderer out so that there
can be different renderer implementations for different GL versions or
Direct3D or whatever, but we actually only have one renderer (for GL2) and
we already have OpenGL-specific extension checking code before the renderer
is created, so screw it.  Please feel free to clean this up again later.