Ensure adverts that are open in ChatForms can't be removed, closes #2732
Attempts to remove an advert that is open in a chat form cause it to be
removed when the chat is closed.

Detect when a ChatForm is closed, and call its OnClosed method.

This unlocks any locked adverts using a lock count to allow multiple
chat windows to lock an advert.