Do not assume that pickle leaves the stack content intact
When iterating through a table, the pickler serializes the key and
value directly, assuming that the recursive calls won't replace the
values on the stack. This is WRONG, as it does replace the values in a
number of cases, for instance when serializing only part of an object.
The solution is to copy the things to get pickled before running the
recursion.