ServerAgent class
The idea is simple - make a call to a remote method with some JSON data,
call a success function with some JSON data when its done, or call a
fail function.

This is deliberately not a generic HTTP client.

Two flavours here: the real thing that makes HTTP calls, and the Null
one which just does nothing