Changed the way planner works
Now the planner uses a fixed point to evaluate parameters and the final
orbit. In this way in is possible to really plan a manouver and apply
thrust until satisfied.

In this commit it is only possible to start the transfer manouver from
the actual position of the player. This will be changed in future
commits.