Added new way to handle planner
Now it is possible to specify when the maneuver will start. It adds new
buttons inside SystemView to change that value.
The system still does not work completely: the speed value at a
specific time of the orbit must be calculated. Soon...

Squash commit: reverted changes to json lang data, except en.json