Fix a small inconsistency in time acceleration ACLs
Before this patch, in the normal case of being in the void without
anyone around, your time acceleration was limited in function of
something that has to do with the angular velocity. I dunno what
exactly, BUT I do know that limit was thrown off for a hard 10x you
couldn't get past in case of another ship being in the immediate
vicinity. Which is good, except that sometimes one shouldn't
even be able to reach the 10x limit in question due to the angular
thingy aforementioned.

This commit thus fixes that.