add new core container widget UI::OverlayStack
An OverlayStack draws its contained widgets on top of each other,
and positions them all at 0,0.

In order to get (mouse) event handling right, or closer to right, the
order of iteration in Container::GetWidgetAt needs to be reversed, so that
it happens in the opposite order to draw order (since the UI is drawn from
far to near).