use the TextureBuilder::UI helper when creating a Drawables::TexturedQuad
This fixes a spurious warning about loading a non-power-of-two texture
(marker_dot.png), by fixing the parameters used for the texture builder.
In particular, TextureBuilder::UI sets extendPot=true which means the
TextureBuilder will automatically pad the texture.

Note that this extra padding is unnecessary on modern graphics chipsets,
but that's a separate bug (the logic in TextureBuilder should be checked
and probably improved to avoid padding textures when running on graphics
hardware that doesn't need it).