remove UI::Image::GetImageSize
I added this method recently thinking it would be useful, but it actually
exposes information (image pixel size) that probably shouldn't be exposed
without good reason. In particular, when PPI can be adjusted it's important
to distinguish between the 'pixel dimensions' of the image and the 'display
dimensions' of the image. A method like 'GetImageSize' (returning pixel
dimensions) lacks that distinction.

Anyway, we might want it back in the future but since it isn't used yet
let's take it out until it's really needed.