allow more control over the size of UI::Icon widgets
By default these widgets will take on the height of their container (scaling
the icon uniformly to fit). However, sometimes you want an icon of a particular
size, so I've added a SetSize() method.

Note that when the icon is scaled to fit the height of its container, it
still keeps its full width as returned by CalcLayoutContribution. This means
that if you have, eg:

    HBox():PackEnd({Icon, Label('Hello, world!'):SetFont('SMALL')})

Then the Icon will end up at the same height as the label (as expected), but
the HBox will still give the Icon its full 48 pixels of width instead of the
width that it *really* needs after scaling down to fit the text line height.
This is probably not what you expect or want.

It is fixable by using the two-phase layout system that @robn implemented
for the MultiLineText widget, but that is left for some future commit.