replace shell+perl canonicalise-translations.sh with a Python script
The Python script is faster, because it processes all the files with a single
process rather than spawning a new perl process for each translation JSON file.

Of course, it could have been written in perl and got the same speedup, but I
don't know perl so Python was easier for me.