Equipment: check for commodity copies when saving in dmode
This patch should make the game throw an assertion error any time
someone tries to serialize a copy of a commodity instead of the actual
commodity. Note that the actual check isn't necessarily correct since
one can easily imagine someone recycling the l10n data of another
object, but this is ugly as hell so we should strive to kick that kind
of code out of the main repo. Let the modders do nasty things if they
want, it'll explode in their faces only when they'll test for it (and
the standard user shouldn't get bitten by it since mods cannot activate
dmode, so the user would have used the console, which makes her not so
standard after all).

Poke @impaktor