Clean up the police tables on ship destruction
It was not really possible to make the whole module stateless.
Furthermore, I kind of like the idea of doing that in our own modules,
since it shows potential modders (and our future selves) that THOU SHALT
SUBSCRIBE TO onShipDestroyed* FOR THINE SHIP REFERENCES. Nanmého.

(*) OK, and maybe the one for the ships leaving the system, depending on
the context

Fixes #3512