Add Language/culture specific names to name gen
This moves name generation to data/cultures/, to make first and last name
match the language, as well as sex of first and last name (e.g. for Russian
and Greek). Old name gen is now a subset, as a "misc" language.

This also brings contributor names up to date in to misc.