Add a TCP Lua console.
This patch adds a new configure option --enable-remote-lua-repl. If that
is given, When starting a game (not in the menu), pioneer listens to
port 12345 (can be changed with -DREMOTE_LUA_REPL_PORT=...). Connect to
it (telnet 127.0.0.1 12345), and you can evaluate expressions like in
the in-game console. This can be used to interface an external lua IDE
to a running game.

At the moment, the code only works on Linux, but the changes for Windows
should be small (mostly some header files). This is only relevant if
--enable-remote-lua-repl is given when configuring!