Add an "advice" module, giving useful information on the BBS.
For now, the advice/rumour/tale is specific to the station. Thinking is: on
that particular station someone lives, or some meme/information/rumour has
taken hold, and in the spirit of sharing this is freely posted on the BBS.

The advice is some "lore" or information that adds useful information
on how to play the game, or things to discover, or make use of in game play.

Ideally, this information would "leak" from some other more relaxed social
setting, like a bar, from more experienced pilots / bartender.