Add support for AppVeyor
AppVeyor is a Contiguous Integration (CI) for Windows.

This set of scripts allows to create a Windows build of Pioneer every
time a pull request is submitted or a commit is pushed, in order to make
sure that the code still compiles on Windows.

It also makes it much easier for the developers to create release builds
for Windows, since they can just redistribute the setup.exe downloaded
from AppVeyor that corresponds to the release tag.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>