Save as gzip-format file.
Doesn't change save version. Should still be able to load the previous
format. Looks for a GZip header when loading the game to determine
whether it is new or old format.

This just adds a GZip header to the save output by default. This makes
it possible to decompress using 'gzip' on the command line instead of
needing a custom tool.