Introducing more attributes
for hyperdrives:
attribute power, replaces "625.0 *(self.capabilities.hyperclass^2)" in the jumprange calculation
attribute maxfuek, replaces "self.capabilities.hyperclass^2" in the fueluse calculations

initial values are set to mimic existing conditions, nothing in ranges or fuel use has changed, but can now be tuned on an individual drive basis

for all equipment:
attribute stockmod, used in the station update to determine how much of each piece of equipment to stock. will make it easier to make individual peices of equipment more likely to go out of stock.

for example, if stock is 5, stockmod is 4, techlevel is 12, then a techlevel 12 station would have about a 1 in 10 chance or so of having a single item in stock, without stockmod, the station would have most likely 2 or 3 with the chance of being out of stock being about 1 in 10. (im not a mathematician or statistician so numbers are ~~ )