New calculations for station commodity stocks
Major imports should frequently be out of stock,
minor imports low or out of stock,
minor exports always in stock,
and major exports always plentiful stock
hydrogen should always be plentiful in stock regardless.

on average, a station will keep stock to a value of up to 100,000 for each commodity.