Fix bundled vs. external GLEW build configuration options
The normal system include path (on Linux) for GLEW is
/usr/include/GL/glew.h (and the other headers). Previously the code here
has included glew/glew.h, which is incompatible with using the system
GLEW library.
Instead, we need to include GL/glew.h as expected and change the include
path in the build commands as needed depending on whether the system
directory or the bundled GLEW is being used.

It's a little tricky to ensure that we actually get the header from the
correct place for both build configurations, because when we're using
the bundled GLEW configuration /usr/include/ (or the standard include
path) will still be on the include path, so *both* the system glew.h
and the bundled glew.h will be available in the search path. We
just have to rely on the include flag order being right and the compiler
preferring include paths from flags to system include paths.

It should be fine... but it's a pain to actually test that it's correct.