Delete almost all XCode related files.
Phil's OS X build process actually uses the autotools build; this is
automated in scripts/build-osx.sh. Only a couple of files from the osx/
directory appear to be referenced, we can just keep these around.

The X Code project is difficult to keep up to date as we have few OS X
developers; best to get rid of it.