Fix orbit calculation for ships with low speeds
1. Improve numerical method for Kepler's equation
It turned out that numerical methods are implemented with a fixed number of
iterations (5), and they are not always enough. Moreover, testing showed
that in some cases the solution does not converge and even 10,000
iterations are not enough. Therefore, a more slowly converging
bisection method was added, in case the main method fails in 10 iterations.

2. Stabilize low speed orbits
When the ship has low speed, its orbit is a very elongated ellipse,
and its true anomaly is very very close to pi. Trimming the variable to
1 - 1e-6 before taking the arccos does not allow it to get close to pi
enough, so there is a strong shift from the true position. Therefore,
it was removed.