Fix calculation of the orientation matrix and parabolic orbit problem
1. Fix orientation matrix calculation
It was found that the orbit orientation matrix is calculated through
the Euler angles, which are calculated from the velocity and position
vectors, and many divisions are made in which the denominator can
unpredictably approach 0. All this was removed since the matrix is
easily calculated directly from the vectors.

2. Transition without distortion from an elliptical orbit to a hyperbolic
When moving from an elliptical orbit to a hyperbolic one, the eccentricity
passes through 1 (parabola), when the average anomaly is zero everywhere,
except for infinity. Therefore, the ship at this time is attracted to
the point of the pericenter, because there is not enough accuracy to
go from the true anomaly to the middle anomaly and back, and stay in
the same coordinates. Therefore, I suggest jumping from an ellipse
with e = 0.9999 to a hyperbole with e = 1.0001. These orbits are
very similar, almost identical. Moreover, ellipses with
0.9999 < e < 1 are also considered as hyperbolas, so that there are
no distortions at a low speed of the ship.