Small additional improvements to orbit calculation
Was: orbit with 0.9999 < e < 1.0001 forced to e = 1.0001 (hyperbola)
Now: orbit with 0.9999 < e < 1      forced to e = 0.9999 (ellipse)
     orbit with 1   <=   e < 1.0001 forced to e = 1.0001 (hyperbola)

In an elliptical orbit, when the eccentricity is fixed at 0.9999, and
the speed increases, the semi-major axis shrinks and the apocenter moves
toward the star. Therefore, a condition check has been added that prevents
the ship from shifting in this case.