Minor improvements to System Map
1. Reject back-projected 2d objects.
Since the function Gui::Screen::Project does not cull points
(and most likely should not), this must be done manually, as was
done in the SectorView.cpp for labels. Checking z < 1 in the Normalized Device
Coordinates discards points that are further than the zfar clipping plane,
and beyond the center of the camera. The points between the center of the
camera and the znear clipping plane remain. They may be needed.

2. Adjust z-far plane for projection to clip space.
With a severe zooming in, the coordinates of objects can become too large and
fly out of the z-far plane. Therefore, it was done so that the z-far plane
moved away proportionally. But with a severe zooming out, it z-far plane can
become smaller than DEFAULT_VIEW_DISTANCE, and this is also taken into account.

3. Change the axis of rotation when dragging the mouse horizontally.
Set horizontal rotation around the axis of the ecliptic. It seems to me that
it’s more convenient to examine the system. For good, it was necessary to
rename the variable from m_rot_z to m_rot_y, but I did not want to
recompile everything.

4. Do not calculate and draw the player’s orbit when it is docked or landed.