New require() implementation
We now load files via Lua 5.2 require() semantics instead of raw paths.
This means we've dropped the current directory implicit lookup, and
instead we use module names to semi-uniquely identify a package.

We mimic the behavior of package.path for the data/libs/ directory,
with an eye to allowing client code to extend that behavior later on.

Loaded modules are cached in a two-stage cache; the first stage maps
module names to file paths to avoid duplicate loads, and the second
stage maps file paths to the return value of a module.

package.reimport(name) can be used to purge the cache of a specific
name and force it to be reimported (as the name suggests). It returns
the imported module, similarly to require().

This implementation deprecates import() and translates old-style calls
to that function to the new module names. It complains loudly in the log
when it does so, so we can keep track of what modules need to be updated.