Make police non-persistent in Goodstrader
Idea: one should be able to recognize the trustworthy black market merchants,
by them being persistent to each station. So you learn to recognize the name
of the market, e.g. "Honest Jim".

However, if _both_ police and the _true_ black market merchant are persistent,
you will recognize the name of both, so now you need to remember _which_ was
the fake one.

I much prefer the way it was in Frontier: the undercover police had a new name
every time you docked, so when returning to a station you would only recognize
one of the trader-names.