Prevent unwanted font face changes
When a character appears in the text, that is not in the default font
face, it is looked up in the spare faces of the font. Moreover, the
default range also began to be baked from the spare face. I made that
only characters that are not in the default face are baked from the
spare face.

Also renamed ImFontAtlas::GlyphRangesBuilder to ImFontGlyphRangesBuilder
because:
contrib/imgui/imgui.h:2181:
typedef ImFontGlyphRangesBuilder GlyphRangesBuilder; // OBSOLETED in 1.67+